OPHJOD-1657: Update Footer
